+++ CHANGELOG.md
@@ -12,6 +12,7 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 ### Changed
 
 - Changed resource display name in notifications.
+- Fixed remove from cache table on booking delete.
 
 ## [1.2.0] - 2023-11-15